Tue, Nov 21, 2023, 6:46 PM +05:30

+++ week1/calc.c
@@ -5,7 +5,7 @@ int main()
 {
     long a= get_int("value of a?\n");
     long b=get_int("value of b?\n");
-    long c=a+b;
-    printf("add=%ld\n",c);
+    double c=a+b;
+    printf("add=%f\n",c);
 }
